+++ SuffixArrayModule/src/main/com/util/Operator.java
@@ -5,5 +5,5 @@ package com.util;
 */
 public interface Operator
 {
-    public int apply(int leftValue, int rightValue);
+    public double apply(double leftValue, double rightValue);
 }
