+++ SuffixArrayModule/src/main/com/util/Operators.java
@@ -7,28 +7,28 @@ public class Operators {
 
     public static final Operator ADD_OPERATOR = new Operator() {
         @Override
-        public int apply(int leftValue, int rightValue) {
+        public double apply(double leftValue, double rightValue) {
             return leftValue+rightValue;
         }
     };
 
     public static final Operator SUBTRACT_OPERATOR = new Operator() {
         @Override
-        public int apply(int leftValue, int rightValue) {
+        public double apply(double leftValue, double rightValue) {
             return leftValue-rightValue;
         }
     };
 
     public static final Operator MULTIPLY_OPERATOR = new Operator() {
         @Override
-        public int apply(int leftValue, int rightValue) {
+        public double apply(double leftValue, double rightValue) {
             return leftValue*rightValue;
         }
     };
 
     public static final Operator DIVIDE_OPERATOR = new Operator() {
         @Override
-        public int apply(int leftValue, int rightValue) {
+        public double apply(double leftValue, double rightValue) {
             return leftValue/rightValue;
         }
     };
